+++ app/models/user.rb
@@ -698,7 +698,6 @@ class User < ApplicationRecord
 
   def initialize_attributes
     self.new_post_navigation_layout = true
-    self.enable_auto_complete = true
     self.always_resize_images = true
   end