+++ app/policies/user_policy.rb
@@ -45,11 +45,10 @@ class UserPolicy < ApplicationPolicy
       :blacklisted_tags, :time_zone, :per_page, :custom_style, :theme,
       :receive_email_notifications, :always_resize_images,
       :new_post_navigation_layout, :enable_private_favorites,
-      :hide_deleted_posts, :style_usernames, :enable_auto_complete,
-      :show_deleted_children, :disable_categorized_saved_searches,
-      :disable_tagged_filenames, :disable_cropped_thumbnails,
-      :disable_mobile_gestures, :enable_safe_mode, :enable_desktop_mode,
-      :disable_post_tooltips,
+      :hide_deleted_posts, :style_usernames, :show_deleted_children,
+      :disable_categorized_saved_searches, :disable_tagged_filenames,
+      :disable_cropped_thumbnails, :disable_mobile_gestures, :enable_safe_mode,
+      :enable_desktop_mode, :disable_post_tooltips,
     ].compact
   end